requirements.txt
langchain
python-dotenv
langchain-community
google-generativeai
streamlit
